исправил условие для 3 задания
